feat: sub-pixel registration (parabolic peak + bilinear shift)
phase_correlate rounded the alignment to the nearest whole pixel — up to a ½-pixel error per
frame, which softens the stack. Refine the correlation peak with a parabolic fit against its
(wrapping) neighbours along each axis, returning a sub-pixel (f32, f32) shift; shift_image now
resolves fractional shifts by bilinear interpolation (an integer shift still reproduces the exact
copy). Net: tighter registration, sharper result.

Ripples: phase_correlate/shift_image are f32 now; the pipeline passes them through unchanged;
synthetic jitter casts to f32. New test proves a 5.4-px shift is recovered to <0.3 px (vs the
0.4 px error nearest-integer rounding would leave). Core tests 6/6, release builds clean.

+++ core/src/synthetic.rs
@@ -49,7 +49,7 @@ pub fn capture(truth: &Gray, n: usize, seed: u64) -> Vec<Frame> {
     for _ in 0..n {
         let sx = rng.gen_range(-6..=6);
         let sy = rng.gen_range(-6..=6);
-        let shifted = align::shift_image(truth, sx, sy);
+        let shifted = align::shift_image(truth, sx as f32, sy as f32);
 
         // Variable seeing: most frames soft, a few sharp.
         let sigma = rng.gen_range(0.4f32..3.0);
